IOS-2354 Separate on files
